requirements.txt
django
django-model-utils
markdown2
django-adminlte-admin
pygments